move pydantic models to new file
