+++ api/pydantic_models.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+from datetime import datetime
+
+
+class TaskCreate(BaseModel):
+    title: str
+    due_to: datetime
+
+class UserModel(BaseModel):
+    username: str
+    password: str
