+++ api/routers/users.py
@@ -1,18 +1,15 @@
 from fastapi import APIRouter, Depends
-from pydantic import BaseModel
+from ..pydantic_models import UserModel
 from sqlalchemy.ext.asyncio import AsyncSession
 from db.session import get_db_session
 from core.user_service import new_user, login_user
 from ..security import create_access_token
 
+
 router = APIRouter(
     prefix="/users",
     tags=["users"]
 )
-
-class UserModel(BaseModel):
-    username: str
-    password: str
 
 @router.post("/register")
 async def register(user: UserModel, session: AsyncSession = Depends(get_db_session)):
